refactor raptor dispatchers to worker base class
