+++ examples/misc/raptor_worker.py
@@ -4,35 +4,16 @@ import random
 
 import radical.pilot as rp
 
+RANKS = 1
+
 
 # ------------------------------------------------------------------------------
 #
-class MyWorker(rp.raptor.MPIWorker):
+class MyWorker(rp.raptor.DefaultWorker):
     '''
     This class provides the required functionality to execute work requests.
     In this simple example, the worker only implements a single call: `hello`.
     '''
-
-    class MyMPIWorkerRank(rp.raptor.MPIWorkerRank):
-
-        def __init__(self, rank_task_q_get, rank_result_q_put, event,
-                           log, prof, base):
-            super().__init__(rank_task_q_get, rank_result_q_put, event,
-                             log, prof, base)
-
-            self.register_mode('foo', self._dispatch_foo)
-
-
-        def _dispatch_foo(self, task):
-
-            import pprint
-            self._log.debug('==== running foo\n%s',
-                    pprint.pformat(task['description']))
-
-            return 'out', 'err', 0, None, None
-
-
-
 
     # --------------------------------------------------------------------------
     #
@@ -40,11 +21,18 @@ class MyWorker(rp.raptor.MPIWorker):
 
         super().__init__(cfg)
 
+        self.register_mode('foo', self._dispatch_foo)
+
+
     # --------------------------------------------------------------------------
     #
-    def get_rank_worker(self):
+    def _dispatch_foo(self, task):
 
-        return self.MyMPIWorkerRank
+        import pprint
+        self._log.debug('==== running foo\n%s',
+                pprint.pformat(task['description']))
+
+        return 'out', 'err', 0, None, None
 
 
     # --------------------------------------------------------------------------
@@ -67,7 +55,7 @@ class MyWorker(rp.raptor.MPIWorker):
         td = rp.TaskDescription({
                 'mode'            : rp.TASK_EXECUTABLE,
                 'scheduler'       : None,
-                'ranks'           : 2,
+                'ranks'           : RANKS,
                 'executable'      : '/bin/sh',
                 'arguments'       : ['-c',
                             'echo "hello $RP_RANK/$RP_RANKS: $RP_TASK_ID"']})
@@ -76,7 +64,7 @@ class MyWorker(rp.raptor.MPIWorker):
               # 'uid'             : 'task.call.w.000000',
               # 'timeout'         : 10,
                 'mode'            : rp.TASK_EXECUTABLE,
-                'ranks'           : 2,
+                'ranks'           : RANKS,
                 'executable'      : 'radical-pilot-hello.sh',
                 'arguments'       : ['1', 'task.call.w.000000']})
