+++ src/radical/pilot/__init__.py
@@ -31,6 +31,8 @@ from .pilot_manager             import PilotManager
 from .pilot                     import Pilot
 from .pilot_description         import PilotDescription
 
+pythontask = PythonTask.pythontask
+
 
 # ------------------------------------------------------------------------------
 # make submodules available -- mostly for internal use
